app/requirements.txt
asgiref==3.6.0
Django==3.0.7
pytz==2022.7.1
sqlparse==0.4.3
psycopg2-binary==2.8.5
gunicorn==20.0.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
